requirements.txt
Django==1.9
django-hstore==1.4.2
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytesseract==0.1.7
pytz==2017.3
